refactor: User 디렉토리 안에 있는 Favorite, Management 디렉토리 분리
