+++ server/module-core/src/main/java/com/momo/domain/favorite/repository/FavoriteGroupCustomRepository.java
@@ -1,6 +1,6 @@
-package com.momo.domain.user.repository;
+package com.momo.domain.favorite.repository;
 
-import com.momo.domain.user.dto.FavoriteGroupCardResponse;
+import com.momo.domain.favorite.dto.FavoriteGroupCardResponse;
 import com.momo.domain.user.entity.User;
 import java.util.List;
