+++ server/module-core/src/test/java/com/momo/domain/group/repository/GroupRepositoryTest.java
@@ -9,7 +9,7 @@ import com.momo.domain.group.dto.GroupResponse;
 import com.momo.domain.group.dto.GroupSearchConditionRequest;
 import com.momo.domain.group.entity.Category;
 import com.momo.domain.group.entity.Group;
-import com.momo.domain.user.entity.FavoriteGroup;
+import com.momo.domain.favorite.entity.FavoriteGroup;
 import com.momo.domain.user.entity.SocialProvider;
 import com.momo.domain.user.entity.User;
 import java.time.LocalDate;
